Auto-import kotlin test in box tests
