+++ compiler-tests/src/test/data/box/aggregation/RepeatedContributesBindingAnvilInteropWorksForBoundTypeAndIgnoreQualifier.kt
@@ -1,7 +1,6 @@
 // WITH_ANVIL
 
 import com.squareup.anvil.annotations.ContributesBinding
-import kotlin.test.*
 
 interface ContributedInterface
 interface SecondInterface
