+++ compiler-tests/src/test/data/box/cycles/BindsCycleGraph.kt
@@ -1,5 +1,3 @@
-import kotlin.test.*
-
 interface Foo
 
 @Inject class Bar(val fooProvider: Provider<Foo>) : Foo
