+++ compiler-tests/src/test/data/box/multibindings/MultibindingGraphWithWithScopedMapProviderDeps.kt
@@ -1,5 +1,3 @@
-import kotlin.test.*
-
 @SingleIn(AppScope::class)
 @DependencyGraph
 abstract class MultibindingGraphWithWithScopedMapProviderDeps {
